+++ server/src/com/openrsc/server/model/world/region/TileValue.java
@@ -1,11 +1,11 @@
 package com.openrsc.server.model.world.region;
 
 public class TileValue {
-	public int traversalMask = 0;
-	public int diagWallVal = 0;
+	public byte traversalMask = 0;
+	public short diagWallVal = 0;
 	public byte horizontalWallVal = 0;
 	public byte overlay = 0;
 	public byte verticalWallVal = 0;
-	public int elevation = 0;
+	public byte elevation = 0;
 	public boolean projectileAllowed = false;
 }
